backend/requirements.txt
google-adk
python-dotenv
pydantic
pandas
numpy
fastapi
uvicorn
requests